integral.dierckx(DierckxSpline)	R Documentation

Spline Integration

Description

Calculates the integral of a spline function s(x) of degree k, which is given in its normalized b-spline representation

Usage


integral(x, ...)
## S3 method for class 'dierckx':
integral(x, from = NULL, to = NULL, ...)

Arguments

x 	 An object of class dierckx.

from 	 Lower integration bound. If NULL, the minimum knot value is used.

to 	 Upper integration bound. If NULL, the maximum knot value is used.

... 	 ignored

Details

s(x) is considered to be identically zero outside the interval (t(k+1),t(n-k)), where t are the knot values. For this reason, from and to are forced to be in or on the boundaries of the knots.

Value

The value of the integral.

Author(s)

Sundar Dorai-Raj

References

Dierckx, P. (1991) Curve and Surface Fitting with Splines, Oxford Science Publications.

See Also

curfit, deriv.dierckx, spline, smooth.spline

Examples


x <- seq(0, 1, 0.1)
y <- (1 - x)^3
z <- curfit(x, y, method = "ls", knots = seq(0, 1, 0.2))

plot(x, y, type = "p")
lines(x, fitted(z), col = "blue")

integral(z)
integrate(function(x) (1 - x)^3, 0, 1)

integral(z, 0.5, 0.6)
integrate(function(x) (1 - x)^3, 0.5, 0.6)

[Package DierckxSpline version 1.0.0 Index]